+++ src/models/base/MHObject.js
@@ -8,12 +8,13 @@ import MHCache from '../internal/MHCache.js';
 import { MHMetadata } from '../meta/MHMetadata.js';
 import MHSocial from '../social/MHSocial.js';
 import MHUserSocial from '../social/MHUserSocial.js';
+import MHSocialMetrics from '../social/MHSocialMetrics.js';
 
 var childrenConstructors = {};
 var __cachedRootResponses = {};
 
 // Create Cache
-export var mhidLRU = new MHCache(1000);
+export var mhidLRU = new MHCache(3000);
 
 if (typeof window !== 'undefined') {
   if (window.location.host === 'local.mediahound.com:2014') {
@@ -62,7 +63,8 @@ export default class MHObject {
       primaryImage: { mapper: MHObject.create },
       secondaryImage: { mapper: MHObject.create },
       social: MHSocial,
-      userSocial: MHUserSocial
+      userSocial: MHUserSocial,
+      socialMetrics: MHSocialMetrics
     };
   }
 
@@ -118,12 +120,12 @@ export default class MHObject {
         // check cache
         //log('in create function trying to parseArgs: \n\n' , args);
 
-        if (mhidLRU.has(args.metadata.mhid) || mhidLRU.has(args.mhid)) {
+
+
+        var foundObject = mhidLRU.get(args.metadata.mhid);
+        if (foundObject) {
           log('getting from cache in create: ' + args.metadata.mhid);
-          var foundObject = mhidLRU.get(args.metadata.mhid);
-          if (foundObject) {
-            foundObject.mergeWithData(args);
-          }
+          foundObject.mergeWithData(args);
           return foundObject;
         }
 
@@ -410,12 +412,18 @@ export default class MHObject {
     log('in fetchByMhid, looking for: ', mhid, 'with view = ', view);
 
     // Check LRU for mhid
-    if (!force && mhidLRU.has(mhid)) {
-      return Promise.resolve(mhidLRU.get(mhid));
+    if (!force) {
+      const entryFromCache = mhidLRU.get(mhid);
+      if (entryFromCache) {
+        return Promise.resolve(entryFromCache);
+      }
     }
     // Check LRU for altId
-    if (!force && mhidLRU.hasAltId(mhid)) {
-      return Promise.resolve(mhidLRU.getByAltId(mhid));
+    if (!force) {
+      const entryFromCache = mhidLRU.getByAltId(mhid);
+      if (entryFromCache) {
+        return Promise.resolve(entryFromCache);
+      }
     }
 
     var prefix  = MHObject.getPrefixFromMhid(mhid),
@@ -638,17 +646,22 @@ export default class MHObject {
       });
   }
 
-  responseCacheKeyForPath(path) {
-    return '__cached_' + path;
+  responseCacheKeyForPath(path, params) {
+    return '__cached_' + path + '_' + JSON.stringify(params, (k, v) => {
+      if (k === 'access_token') {
+        return undefined;
+      }
+      return v;
+    });;
   }
 
-  cachedResponseForPath(path) {
-    var cacheKey = this.responseCacheKeyForPath(path);
+  cachedResponseForPath(path, params) {
+    var cacheKey = this.responseCacheKeyForPath(path, params);
     return this.cachedResponses[cacheKey];
   }
 
-  setCachedResponse(response, path) {
-    var cacheKey = this.responseCacheKeyForPath(path);
+  setCachedResponse(response, path, params) {
+    var cacheKey = this.responseCacheKeyForPath(path, params);
     this.cachedResponses[cacheKey] = response;
   }
 
@@ -747,7 +760,7 @@ export default class MHObject {
     params.view = view;
 
     if (!force && !next) {
-      var cached = this.cachedResponseForPath(path);
+      var cached = this.cachedResponseForPath(path, params);
       if (cached) {
         return cached;
       }
@@ -781,7 +794,7 @@ export default class MHObject {
     });
 
     if (!next) {
-      this.setCachedResponse(finalPromise, path);
+      this.setCachedResponse(finalPromise, path, params);
     }
 
     return finalPromise;